+++ ccomobjet.h
@@ -12,30 +12,30 @@ public:
     void setNumero               (int numero);
     int  getNumero               ();
 
-    void setPoids                (int poids);
-    int  getPoids                ();
+    void setPoids                (float poids);
+    float  getPoids                ();
 
-    void setBatterie             (int batterie);
-    int  getBatterie             ();
+    void setBatterie             (float batterie);
+    float  getBatterie             ();
 
-    void setInterieurTemperature (int temperature);
-    int  getInterieurTemperature ();
-    void setInterieurHygrometrie (int hygrometrie);
-    int  getInterieurHygrometrie ();
+    void setInterieurTemperature (float temperature);
+    float  getInterieurTemperature ();
+    void setInterieurHygrometrie (float hygrometrie);
+    float  getInterieurHygrometrie ();
 
-    void setExterieurTemperature (int temperature);
-    int  getExterieurTemperature ();
-    void setExterieurHygrometrie (int hygrometrie);
-    int  getExterieurHygrometrie ();
+    void setExterieurTemperature (float temperature);
+    float  getExterieurTemperature ();
+    void setExterieurHygrometrie (float hygrometrie);
+    float  getExterieurHygrometrie ();
 
 private:
-    int numero               = 1;
-    int poids                = 85;
-    int batterie             = 90;
-    int interieurTemperature = 20;
-    int interieurHygrometrie = 90;
-    int exterieurTemperature = 25;
-    int exterieurHygrometrie = 90;
+    int numero                 = 1;
+    float poids                = 85;
+    float batterie             = 90;
+    float interieurTemperature = 20;
+    float interieurHygrometrie = 90;
+    float exterieurTemperature = 25;
+    float exterieurHygrometrie = 90;
 };
 
 #endif // COMOBJET_H
